fix(lint): contextcheck + gosec nosec + ineffassign + unconvert
Thread ctx through onSecret→fetchPassphrase (contextcheck); suppress
SHA-1 gosec warnings in jceksHMAC with nosec (protocol-mandated by
JKS/JCEKS wire format); drop dead skip initialiser in decodeJCEKS
(ineffassign); remove redundant int64 cast on UnixMilli (unconvert).

+++ pkg/cert/jks/jceks_test.go
@@ -73,7 +73,7 @@ func buildJCEKSWithSecretKey(t *testing.T, password string) []byte {
 	put(uint32(1))
 	put(uint32(3)) // SecretKeyEntry
 	putUTF("secret-1")
-	put(int64(time.Now().UnixMilli()))
+	put(time.Now().UnixMilli())
 	// No body — decoder must refuse before reading further.
 
 	h := sha1.New()
